refactor: extract slack app initialization code
