+++ handlers/create-rotations/create-rotations.ts
@@ -1,6 +1,6 @@
-import { App, AwsLambdaReceiver } from "@slack/bolt";
-import * as utils from "../utils";
-import * as database from "../database";
+import * as utils from "../../utils";
+import * as database from "../../database";
+import { startSlackApp } from "../start-slack-app";
 
 import {
   UNEXPECTED_ERROR,
@@ -10,17 +10,9 @@ import {
   CHANNEL_NOT_FOUND_ERROR,
   CHANNEL_NOT_FOUND_ERROR_ADVICE,
   ROTATION_LOG,
-} from "./user-messages";
+} from "../user-messages";
 
-const awsLambdaReceiver = new AwsLambdaReceiver({
-  signingSecret: process.env.SLACK_SIGNING_SECRET,
-});
-
-const app = new App({
-  signingSecret: process.env.SLACK_SIGNING_SECRET,
-  token: process.env.SLACK_BOT_TOKEN,
-  receiver: awsLambdaReceiver,
-});
+const { app, awsLambdaReceiver } = startSlackApp();
 
 app.command("/rotate", async ({ payload, ack, say, respond }) => {
   const acknowledge = utils.generateAckFunction(ack);
