+++ handlers/start-slack-app.ts
@@ -1,0 +1,16 @@
+import { App, AwsLambdaReceiver } from "@slack/bolt";
+
+const awsLambdaReceiver = new AwsLambdaReceiver({
+  signingSecret: process.env.SLACK_SIGNING_SECRET,
+});
+
+const app = new App({
+  signingSecret: process.env.SLACK_SIGNING_SECRET,
+  token: process.env.SLACK_BOT_TOKEN,
+  receiver: awsLambdaReceiver,
+});
+
+export const startSlackApp = () => ({
+  app,
+  awsLambdaReceiver,
+});
